bootstrap wget
Currently apt-cyg has a circular dependency on wget. This can be resolved the
following ways, but each has problems

- setup-x86_64
  you can run a command like "setup-x86_64 -qP wget" but this is not desirable
  because it just spawns the gui, which was probably what apt-cyg was designed
  to prevent in the first place

- using /dev/tcp
  this is a good solution but does not work for "https://" URLs

- using telnet
  this is a windows component but is disabled by default

- using curl
  like wget, this does not come with base Cygwin

- using powershell
  this is available for XP and later, but only comes preinstalled with Windows 7
  and later

This commit would allow a "wget" Bash function to be used if "wget.exe" is not
found, for example
http://github.com/svnpenn/dotfiles/blob/19b7b76/.bashrc#L54-L67
My function uses powershell, but a function using any of the above methods would
work.